Add files via upload
The website for Lingfield Flowers Club is now in the next stage and in preparation for publishing. The biggest works were on animated advert in the middle of the home page and then on gallery with thumbnails and enlarged pictures selected by users. It was a lot of testing and studying of both CSS and Javascript. There were also adjustments done as per requirements from the flower club management. (For example the gallery was originally much simpler). This version also has media queries so it works and looks good on most smartphones and tablets as well.